Move methods of `JsString` to `JsStr`
